Allow specifying block response in configuration
NXDOMAIN is the traditional response code for blocking DNS queries, but
can make it difficult to differentiate between actual DNS problems and
DNS blocklists. This change introduces the ability to respond with
REFUSED instead, to make it easier to identify DNS blocks.

The default behaviour remains to respond with NXDOMAIN, so no additional
configuration is required to maintain current behaviour.